// front-end/src/app/api/generated/api.ts
export interface components {
    schemas: {
// ...
        ReviewDTO: {
            /** Format: int32 */
            id?: number;
            patientId?: string | null;
            doctorId?: string | null;
            isReviewAnonymous?: boolean | null;
            /** Format: int32 */
            waitTimeRating?: number | null;
            /** Format: int32 */
            bedsideMannerRating?: number | null;
            /** Format: int32 */
            overallRating?: number | null;
            review?: string | null;
            isDoctorRecommended?: boolean | null;
            /** Format: date-time */
            reviewDate?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}